Moved suggest_username into utils.py
