+++ kitsune/users/utils.py
@@ -1,3 +1,4 @@
+import bisect
 import logging
 from smtplib import SMTPException
 
@@ -95,3 +96,32 @@ def add_to_contributors(request, user):
         return mail
 
     email_utils.send_messages([_make_mail(request.LANGUAGE_CODE)])
+
+
+def suggest_username(email):
+    username = email.split('@', 1)[0]
+
+    username_regex = r'^{0}[0-9]*$'.format(username)
+    users = User.objects.filter(username__iregex=username_regex)
+
+    if users.count() > 0:
+        ids = []
+        for u in users:
+            # get the number at the end
+            i = u.username[len(username):]
+
+            # incase there's no number in the case where just the base is taken
+            if i:
+                i = int(i)
+                bisect.insort(ids, i)
+            else:
+                ids.insert(0, 0)
+
+        for index, i in enumerate(ids):
+            if index + 1 < len(ids):
+                if i + 1 != ids[index + 1]:
+                    break
+
+        username = '{0}{1}'.format(username, i + 1)
+
+    return username
